Refatoração completa do projeto e melhoria da documentação
Este commit inclui uma série de melhorias na organização do projeto,
clareza do código e robustez da documentação.

Principais alterações realizadas:

1.  Reorganização da Estrutura de Arquivos:
    *   Criados diretórios dedicados: `config/` para templates de
        configuração, `models/` para modelos de ML e `src/` para o script
        Python principal.
    *   Arquivos JavaScript da interface web movidos para
        `sistema_de_monitoramento/static/js/`.
    *   Templates de configuração renomeados com sufixo `.example`.

2.  Melhoria da Documentação (README.md):
    *   Atualizada para refletir a nova estrutura de arquivos.
    *   Instruções de configuração e instalação detalhadas, incluindo
        a adição da dependência `pandas`.
    *   Clarificadas as etapas para configurar Firebase (backend e frontend),
        Telegram, e o script de detecção.
    *   Melhoradas as instruções para deploy da interface web no Firebase
        Hosting, especialmente sobre o manuseio do `firebaseConfig.js`.

3.  Refatoração do Código Python (`src/deteccao.py.example`):
    *   Configurações (caminhos de arquivo, tokens, parâmetros de detecção)
        movidas para o topo do script, com suporte para configuração via
        variáveis de ambiente e valores padrão.
    *   Melhorado o tratamento de erros e a robustez geral do script.
    *   Código reorganizado em funções mais claras (inicialização de
        Firebase, envio de notificações, stream de vídeo).
    *   Lógica de alerta e detecção aprimorada.

4.  Configuração do Firebase:
    *   `config/firebaseConfig.example.js` ajustado para exportar
        corretamente as instâncias `db` e `auth` do Firebase.
    *   Caminhos de importação de `firebaseConfig.js` atualizados nos
        arquivos HTML e JavaScript da interface web para uma estratégia
        consistente de desenvolvimento local e deploy.

5.  Atualização do `.gitignore`:
    *   Revisado e atualizado para ignorar corretamente os arquivos
        gerados pelo usuário a partir dos templates (ex: `config/firebase_key.json`,
        `config/firebaseConfig.js`, `src/deteccao.py`).
    *   Adicionada regra para ignorar modelos PyTorch grandes (`models/*.pt`).

6.  Testes Manuais (Simulados):
    *   Processo de configuração e execução foi simulado com base no
        README atualizado, levando a ajustes nos caminhos de importação e
        na lista de dependências.

O objetivo geral foi tornar o projeto mais fácil de entender, configurar,
manter e usar.
