+++ sistema_de_monitoramento/historico.html
@@ -198,8 +198,9 @@
     <!-- Sidebar Overlay for Mobile -->
     <div id="sidebarOverlay" class="fixed inset-0 bg-black bg-opacity-50 z-30 lg:hidden hidden"></div>
 
-    <script type="module" src="firebaseConfig.js"></script>
-    <script type="module" src="carregarTabela.js"></script>
+    <script type="module" src="firebaseConfig.js"></script> <!-- Assumindo que está na raiz de sistema_de_monitoramento/ para deploy -->
+    <script type="module" src="static/js/carregarTabela.js"></script>
+    <script type="module" src="static/js/carregarImagem.js"></script>
     <script>
         // Sidebar toggle functionality
         const sidebar = document.getElementById('sidebar');
