+++ sistema_de_monitoramento/login.html
@@ -166,6 +166,6 @@
         </div>
     </div>
 
-    <script type="module" src="back.js"></script>
+    <script type="module" src="static/js/back.js"></script>
 </body>
 </html>
