+++ sistema_de_monitoramento/usuarios.html
@@ -240,7 +240,8 @@
     <!-- Sidebar Overlay for Mobile -->
     <div id="sidebarOverlay" class="fixed inset-0 bg-black bg-opacity-50 z-30 lg:hidden hidden"></div>
 
-    <script type="module" src="userManagement.js"></script>
+    <script type="module" src="firebaseConfig.js"></script> <!-- Assumindo que está na raiz de sistema_de_monitoramento/ para deploy -->
+    <script type="module" src="static/js/userManagement.js"></script>
     <script>
         // Sidebar toggle functionality
         const sidebar = document.getElementById('sidebar');
